requirements.txt
fastapi[standard]
psycopg[binary]
PyJWT
python-dotenv
email-validator
pydantic
